requirements.txt
langgraph>=0.2.35,<0.4
langchain-aws>=0.2.1,<0.3
python-dotenv>=1.0,<2.0
pandas>=2.2.3,<2.3
sqlalchemy>=2.0.35,<2.1
psycopg2-binary>=2.9.9,<3.0
pydantic>=2.10,<3.0
boto3>=1.35.25,<2.0
openpyxl>=3.1.5,<4.0
lxml>=5.0,<6.0
html5lib>=1.1,<2.0
streamlit>=1.39,<2.0
# LLM-focused pipeline: no crawl4ai dependency in base install.
# Optional LLM providers — install only what you use
langchain-openai>=0.2,<0.4
langchain-google-genai>=2.0,<3.0
langchain-groq>=0.2,<0.4
langchain-ollama>=0.2,<0.4
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
strands-agents-builder>=0.1.10
bedrock-agentcore>=1.4.8
bedrock-agentcore-starter-toolkit>=0.3.3
playwright>=1.58.0
